test(django-cf): restore the original filenames of the migrated suites
The in-worker migration merged six host test modules into two, and renamed a
lot of what it moved, which made it impossible to check the port against main
one test at a time. Split them back to the names main used:

  test_db.py             -> test_base_engine.py, test_d1_backend.py,
                            test_do_backend.py
  test_middleware_wsgi.py -> test_cloudflare_access.py, test_storage_errors.py,
                            test_wsgi_handler.py

Bodies are moved verbatim; only files, class names and test names change. Each
file carries its own helpers rather than importing a shared module, matching
how the originals looked. Filenames are the suite names the in-worker runner
dispatches on, so this also splits the run into six /run-tests/<suite> calls
per config instead of two, and generated host-side classes are now CamelCase
(TestBaseEngine, not TestBASE_ENGINE).

159 tests on main, 156 now. test_base_engine.py matches main exactly, 68 for
68. The rest:

- 4 source-inspection tests are gone, and should be. They asserted on the text
  of the implementation - that run_query uses `except Exception`, that
  Error.stackTraceLimit is set, that storage.py contains an import. None
  survive contact with a real runtime and none tested behaviour.
- 4 R2 tests that mocked a raising bucket are gone, replaced by real-binding
  equivalents against real missing objects: *_on_exception became
  *_on_not_found. Same intent, different mechanism, so the names changed.
- 2 host-only tests are gone: get_bucket_raises_on_non_worker cannot fail that
  way inside a worker, and the two WSGI header-transformation tests inspected
  handle_wsgi internals the Worker boundary tests now cover end to end.
- 3 genuinely uncovered: size when metadata has no size attribute, save
  preserving content_type, and the mock-only read-raises path.
- 7 added: real D1 read/write round trips, DO storage wiring, and Access user
  provisioning, none of which had host-side counterparts.

Full per-test mapping in /tmp/mig/report_db.md and /tmp/mig/report_middleware.md
(not committed).

348 passed, 2 skipped.
