+++ packages/django-cf/tests/in_worker/worker/src/test_base_engine.py
@@ -1,102 +1,11 @@
-"""Database backend tests executed inside workerd."""
+"""Base engine tests executed inside workerd."""
 
 # pyright: reportMissingImports=false
 
 from decimal import Decimal
-from types import SimpleNamespace
 from unittest.mock import MagicMock, patch
 
 import pytest
-import workers
-
-
-def make_d1_wrapper(defer_foreign_keys=False):
-    from django_cf.db.backends.d1.base import DatabaseWrapper
-
-    wrapper = DatabaseWrapper.__new__(DatabaseWrapper)
-    cursor_state = SimpleNamespace(_defer_foreign_keys=defer_foreign_keys)
-    wrapper.cursor = lambda: cursor_state
-    wrapper.binding = "DB"
-    wrapper.run_sync = lambda value: value
-    return wrapper
-
-
-def make_do_wrapper(defer_foreign_keys=False):
-    from django_cf.db.backends.do.base import DatabaseWrapper
-
-    wrapper = DatabaseWrapper.__new__(DatabaseWrapper)
-    cursor_state = SimpleNamespace(_defer_foreign_keys=defer_foreign_keys)
-    wrapper.cursor = lambda: cursor_state
-    return wrapper
-
-
-class FakeD1Statement:
-    def __init__(
-        self, *, raw_result=None, all_result=None, raw_error=None, all_error=None
-    ):
-        self.raw_result = raw_result or []
-        self.all_result = all_result or {"results": [], "meta": {}}
-        self.raw_error = raw_error
-        self.all_error = all_error
-        self.bound_params = None
-
-    def bind(self, *params):
-        self.bound_params = params
-        return self
-
-    def raw(self):
-        if self.raw_error is not None:
-            raise self.raw_error
-        return self.raw_result
-
-    def all(self):
-        if self.all_error is not None:
-            raise self.all_error
-        return self.all_result
-
-
-class FakeD1Binding:
-    def __init__(self, statement):
-        self.statement = statement
-        self.prepared_queries = []
-
-    def prepare(self, query):
-        self.prepared_queries.append(query)
-        return self.statement
-
-
-class FakeDOArray:
-    def __init__(self, data):
-        self.data = data
-
-    def toArray(self):
-        return self
-
-    def to_py(self):
-        return self.data
-
-
-class FakeDOStatement:
-    def __init__(self, data, *, rows_read=0, rows_written=0):
-        self.data = data
-        self.rowsRead = rows_read
-        self.rowsWritten = rows_written
-
-    def raw(self):
-        return FakeDOArray(self.data)
-
-
-class FakeDOStorage:
-    def __init__(self, statement=None, *, error=None):
-        self.statement = statement
-        self.error = error
-        self.calls = []
-
-    def exec(self, query, *params):
-        if self.error is not None:
-            raise self.error
-        self.calls.append((query, params))
-        return self.statement
 
 
 class TestCFResult:
@@ -694,334 +603,3 @@ class TestCFSQLCompiler:
         )
 
         assert "django_date_trunc('unknown', created_at)" in result
-
-
-class TestD1DatabaseWrapperProcessQuery:
-    def test_process_query_no_params(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "SELECT * FROM users WHERE id = %s", None
-        )
-
-        assert result_query == "SELECT * FROM users WHERE id = ?"
-        assert result_params is None
-
-    def test_process_query_with_params(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "SELECT * FROM users WHERE id = %s AND name = %s", [1, "test"]
-        )
-
-        assert result_query == "SELECT * FROM users WHERE id = ? AND name = ?"
-        assert result_params == [1, "test"]
-
-    def test_process_query_null_param_replaced_with_literal(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "INSERT INTO users (name, email) VALUES (%s, %s)", ["test", None]
-        )
-
-        assert "null" in result_query
-        assert result_params == ["test"]
-
-    def test_process_query_all_null_params(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "INSERT INTO users (name, email) VALUES (%s, %s)", [None, None]
-        )
-
-        assert result_query == "INSERT INTO users (name, email) VALUES (null, null)"
-        assert result_params == []
-
-    def test_process_query_mixed_params(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "UPDATE users SET name = %s, email = %s, age = %s WHERE id = %s",
-            ["test", None, 25, None],
-        )
-
-        assert result_params == ["test", 25]
-        assert result_query.count("?") == 2
-        assert result_query.count("null") == 2
-
-    def test_process_query_with_defer_foreign_keys(self):
-        result = make_d1_wrapper(True).process_query(
-            "INSERT INTO users (name) VALUES (%s)", ["test"]
-        )
-
-        assert "PRAGMA defer_foreign_keys = on" in result
-        assert "PRAGMA defer_foreign_keys = off" in result
-
-    def test_process_query_date_trunc_replacement(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, _ = wrapper.process_query(
-            "SELECT django_date_trunc(%s, created_at, %s, %s) FROM orders",
-            ["year", "UTC", "UTC"],
-        )
-
-        assert "django_date_trunc" not in result_query
-        assert "CASE" in result_query or "STRFTIME" in result_query
-
-
-class TestD1DatabaseWrapperConfiguration:
-    def test_vendor_and_display_name(self):
-        from django_cf.db.backends.d1.base import DatabaseWrapper
-
-        assert DatabaseWrapper.vendor == "cloudflare_d1"
-        assert DatabaseWrapper.display_name == "D1"
-
-
-class TestD1GetConnectionParams:
-    def test_missing_binding_raises_error(self):
-        from django.core.exceptions import ImproperlyConfigured
-
-        wrapper = make_d1_wrapper()
-        wrapper.settings_dict = {"CLOUDFLARE_BINDING": None}
-
-        with pytest.raises(ImproperlyConfigured) as exc_info:
-            wrapper.get_connection_params()
-
-        assert "CLOUDFLARE_BINDING" in str(exc_info.value)
-
-    def test_valid_binding_returns_params(self):
-        wrapper = make_d1_wrapper()
-        wrapper.settings_dict = {"CLOUDFLARE_BINDING": "MY_DB"}
-
-        assert wrapper.get_connection_params() == {"binding": "MY_DB"}
-
-
-class TestD1ExceptionHandling:
-    @pytest.mark.xfail(
-        reason=(
-            "The `except Exception: raise Error(Error.new().stack)` handler in `run_query` swallows this and re-raises a JS Error. Removing that handler is a separate change; these pass once it lands."
-        ),
-        strict=True,
-    )
-    def test_run_query_lets_binding_errors_propagate(self, monkeypatch):
-        wrapper = make_d1_wrapper()
-        error = RuntimeError("binding boom")
-        monkeypatch.setattr(
-            workers,
-            "env",
-            SimpleNamespace(DB=FakeD1Binding(FakeD1Statement(raw_error=error))),
-        )
-
-        with pytest.raises(RuntimeError, match="binding boom"):
-            wrapper.run_query("SELECT * FROM test")
-
-
-class TestD1RunQuery:
-    def test_read_query_returns_rows(self, monkeypatch):
-        wrapper = make_d1_wrapper()
-        statement = FakeD1Statement(raw_result=[[1, "hello"], [2, "world"]])
-        binding = FakeD1Binding(statement)
-        monkeypatch.setattr(workers, "env", SimpleNamespace(DB=binding))
-
-        result = wrapper.run_query("SELECT * FROM test WHERE id = %s", [1])
-
-        assert list(result) == [(1, "hello"), (2, "world")]
-        assert statement.bound_params == (1,)
-        assert binding.prepared_queries == ["SELECT * FROM test WHERE id = ?"]
-
-    def test_write_query_returns_meta(self, monkeypatch):
-        wrapper = make_d1_wrapper()
-        statement = FakeD1Statement(
-            all_result={
-                "results": [["ok"]],
-                "meta": {"rows_read": 1, "rows_written": 2, "last_row_id": 9},
-            }
-        )
-        monkeypatch.setattr(
-            workers, "env", SimpleNamespace(DB=FakeD1Binding(statement))
-        )
-
-        result = wrapper.run_query("INSERT INTO test VALUES (%s)", ["x"])
-
-        assert list(result) == [("ok",)]
-        assert result.rowcount == 2
-        assert result.lastrowid == 9
-
-
-class TestD1ParameterHandling:
-    def test_empty_params_list(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, result_params = wrapper.process_query("SELECT * FROM users", [])
-
-        assert result_query == "SELECT * FROM users"
-        assert result_params == []
-
-    def test_special_characters_in_params(self):
-        wrapper = make_d1_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "INSERT INTO users (name) VALUES (%s)", ["test'; DROP TABLE users; --"]
-        )
-
-        assert result_params == ["test'; DROP TABLE users; --"]
-        assert "?" in result_query
-
-    def test_unicode_params(self):
-        wrapper = make_d1_wrapper()
-
-        _, result_params = wrapper.process_query(
-            "INSERT INTO users (name) VALUES (%s)", [""]
-        )
-
-        assert result_params == [""]
-
-    def test_large_number_of_params(self):
-        wrapper = make_d1_wrapper()
-
-        placeholders = ", ".join(["%s"] * 50)
-        result_query, result_params = wrapper.process_query(
-            f"INSERT INTO test VALUES ({placeholders})", list(range(50))
-        )
-
-        assert result_query.count("?") == 50
-        assert len(result_params) == 50
-
-
-class TestDODatabaseWrapperProcessQuery:
-    def test_process_query_no_params(self):
-        wrapper = make_do_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "SELECT * FROM users WHERE id = %s", None
-        )
-
-        assert result_query == "SELECT * FROM users WHERE id = ?"
-        assert result_params is None
-
-    def test_process_query_with_params(self):
-        wrapper = make_do_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "SELECT * FROM users WHERE id = %s AND name = %s", [1, "test"]
-        )
-
-        assert result_query == "SELECT * FROM users WHERE id = ? AND name = ?"
-        assert result_params == [1, "test"]
-
-    def test_process_query_null_param_replaced_with_literal(self):
-        wrapper = make_do_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "INSERT INTO users (name, email) VALUES (%s, %s)", ["test", None]
-        )
-
-        assert "null" in result_query
-        assert result_params == ["test"]
-
-    def test_process_query_with_defer_foreign_keys(self):
-        result = make_do_wrapper(True).process_query(
-            "INSERT INTO users (name) VALUES (%s)", ["test"]
-        )
-
-        assert "PRAGMA defer_foreign_keys = on" in result
-        assert "PRAGMA defer_foreign_keys = off" in result
-
-
-class TestDODatabaseWrapperConfiguration:
-    def test_vendor_and_display_name(self):
-        from django_cf.db.backends.do.base import DatabaseWrapper
-
-        assert DatabaseWrapper.vendor == "cloudflare_durable_objects"
-        assert DatabaseWrapper.display_name == "DO"
-
-    def test_get_connection_params_returns_empty(self):
-        assert make_do_wrapper().get_connection_params() == {}
-
-
-class TestDOMissingDateTruncBug:
-    def test_do_process_query_missing_date_trunc(self):
-        d1_result, _ = make_d1_wrapper().process_query(
-            "SELECT django_date_trunc(%s, created_at, %s, %s) FROM orders",
-            ["year", "UTC", "UTC"],
-        )
-        do_result, _ = make_do_wrapper().process_query(
-            "SELECT django_date_trunc(%s, created_at, %s, %s) FROM orders",
-            ["year", "UTC", "UTC"],
-        )
-
-        assert "django_date_trunc" not in d1_result
-        assert "django_date_trunc" in do_result
-
-
-class TestDOStorageInitialization:
-    def test_storage_module_exists(self):
-        from django_cf.db.backends.do import storage
-
-        storage.set_storage(None)
-        assert storage.get_storage() is None
-
-    def test_run_query_uses_configured_storage(self):
-        from django_cf.db.backends.do import storage
-
-        fake_storage = FakeDOStorage(FakeDOStatement([[1, "ok"]], rows_read=1))
-        storage.set_storage(fake_storage)
-
-        result = make_do_wrapper().run_query("SELECT * FROM test WHERE id = %s", [1])
-
-        assert list(result) == [(1, "ok")]
-        assert fake_storage.calls == [("SELECT * FROM test WHERE id = ?", (1,))]
-
-
-class TestDOQueryExecution:
-    def test_read_query_uses_raw(self):
-        from django_cf.db.backends.do import storage
-
-        fake_storage = FakeDOStorage(FakeDOStatement([], rows_read=0, rows_written=0))
-        storage.set_storage(fake_storage)
-
-        result = make_do_wrapper().run_query("SELECT * FROM users")
-
-        assert result.fetchall() == []
-
-
-class TestDOExceptionHandling:
-    def test_run_query_lets_binding_errors_propagate(self):
-        from django_cf.db.backends.do import storage
-
-        storage.set_storage(FakeDOStorage(error=RuntimeError("storage boom")))
-
-        with pytest.raises(RuntimeError, match="storage boom"):
-            make_do_wrapper().run_query("SELECT * FROM test")
-
-
-class TestDOParamConversion:
-    def test_boolean_true_not_converted_in_process_query(self):
-        wrapper = make_do_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "INSERT INTO test (active) VALUES (%s)", [True]
-        )
-
-        assert result_query == "INSERT INTO test (active) VALUES (?)"
-        assert result_params == [True]
-
-    def test_multiple_none_params_order_preserved(self):
-        wrapper = make_do_wrapper()
-
-        result_query, result_params = wrapper.process_query(
-            "INSERT INTO test (a, b, c, d) VALUES (%s, %s, %s, %s)",
-            [1, None, 2, None],
-        )
-
-        assert result_params == [1, 2]
-        assert result_query == "INSERT INTO test (a, b, c, d) VALUES (?, null, ?, null)"
-
-
-class TestDOPragmaReturnTypeBug:
-    def test_pragma_returns_string_not_tuple(self):
-        result = make_do_wrapper(True).process_query(
-            "INSERT INTO users (name) VALUES (%s)", ["test"]
-        )
-
-        assert isinstance(result, str)
